pyproject.toml
[tool.poetry]
name = "lrcp"
version = "1.0.0"
description = "Панель управления сервером"
authors = ["lordralinc"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
grpcio = ">=1.45.0,<2.0"
fastapi = ">=0.75.0,<1.0"
tortoise-orm = ">=0.19.0,<1.0"
toml = ">=0.10.2,<1.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pydantic = {extras = ["dotenv"], version = "^1.9.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = ">=0.0.5,<1.0"
pyotp = ">=2.6.0,<3.0"
wsproto = "^1.1.0"
fastapi-socketio = "^0.0.9"


[tool.poetry.dev-dependencies]
mypy = ">=0.910"
types-protobuf = ">=0.1.14"
mypy-protobuf = "^3.2.0"
grpcio-tools = ">=1.45.0,<2.0"
Sphinx = "^4.5.0"
sphinx-rtd-theme = "^1.0.0"


[tool.poetry.scripts]
manage="lrcp.__main__:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
